+++ CHANGELOG.md
@@ -7,6 +7,13 @@ file.
 ### Added
 
 ### Changed
+
+### Removed
+
+## [0.13.1] - 2020-05-25
+### Added
+
+### Changed
 - `frozen`, `locked`, `force-clean` and `ignore-tests` flags are now propagated
 to feature configurations.
 - `exclude` argument for packages is now propagated and any features existing